+++ app/Http/Controllers/HomeController.php
@@ -21,8 +21,10 @@ class HomeController extends Controller
      *
      * @return \Illuminate\Contracts\Support\Renderable
      */
-    public function index()
+    public function index(Request $request)
     {
+
+        //return $request->path();
         return view('home');
         //return view('welcome');
     }
