+++ public/css/style.css
@@ -29,3 +29,11 @@ text-transform: capitalize;
   font-size: 14px;
   font-weight: 400;
 }
+
+.flex1{
+flex:1;
+}
+.flex2{
+flex:2;
+}
+
