+++ resources/js/components/layouts/AdminLayout.vue
@@ -11,9 +11,9 @@
         </div> -->
 <appbar/>
         <div class="main-content">
-            <!-- <router-view /> -->
+          <router-view />
           <!-- <transition name="fade">
-          </transition> -->
+          </transition>-->
         </div>
       </div>
     </div>
@@ -21,19 +21,29 @@
 </template>
 
 <script>
+import sidebar from "../backend/sidebar";
+import appbar from "../../components/backend/topappBar";
 
-import sidebar from '../../components/backend/sidebar'
-import appbar from '../../components/backend/topappBar'
 export default {
-
   components: {
-sidebar,
-appbar
-
+    sidebar,
+    appbar
   }
 };
 </script>
 
-<style>
+<style scoped>
 
+.main-content {
+  margin-top: 80px;
+  padding: 0px 15px;
+}
+
+.fade-enter-active,
+.fade-leave-active {
+  transition: opacity 0.4s;
+}
+.fade-enter, .fade-leave-to /* .fade-leave-active below version 2.1.8 */ {
+  opacity: 0;
+}
 </style>
