+++ routes/web.php
@@ -16,7 +16,16 @@
 }); */
 
 //Auth::routes();
-Route::post('login', 'API\RegisterController@login');
-Route::get('/', 'HomeController@index')->name('home');
-Route::get('/{any}', 'HomeController@index')->where('any','.*');
-//Route::get('/{any}/{any}', 'HomeController@index')->name('home');
+Route::post('/login', 'API\RegisterController@login');
+Route::get('/logout', 'API\RegisterController@logout');
+Route::get('/', 'API\RegisterController@index');
+Route::get('/{any}', 'API\RegisterController@index')->where('any','.*');
+
+
+//Route::get('/{any}/{any}', 'API\RegisterController@index');
+
+
+//Route::get('/', 'HomeController@index')->name('home');
+
+//Route::get('/{any}', 'HomeController@index')->where('any','.*');
+// Route::get('/{any}/{any}', 'HomeController@index')->name('home');
